requirements.txt
azure-identity==1.1.0
azure-keyvault-secrets==4.0.0
pytest
